fix 500 error due to missing user agent
This change fixes the logic that checks the request user agent to see if
the browser is unsupported. The current logic assumes that there'll
always be an `HTTP_USER_AGENT` header, which is not always the case. The
new logic makes this logic more flexible.

This also adds some Python unit tests to verify this behavior and adds
instructions on running them to the README.

Fixes #128.