Bump Django support (1.11); install deps w/setup.py
The Django component to this repo has now been tested with 1.11.14, and
the repo remains compatible with Django 1.8.15. We'll assume that means
it is also compatible with Django versions in between those two.

This change is being made in preparation for upgrading cfgov-refresh to
Django 1.11.

At the same time, we're also switching away from requirements.txt to
specify this app's requirements -- in favor of setup.py. This matches
the preferred configuration of other cfgov-refresh satellite apps and
allows for a better install experience. See

  https://www.b-list.org/weblog/2018/apr/25/lets-talk-about-packages/

for more rationale.