deprecate automatic tagging
This change removes the automatic tagging behavior that currently runs
on Travis after every push/merge to master. This makes the behavior of
this repository more consistent with other cfgov-refresh satellite apps:
on tags, Travis will upload the built wheel file to the tag's GitHub
Release.

It also adds a command to .travis.yml that reverts any local changes
made to package-lock.json when 'npm install' is run. This is because
package-lock.json might be modified due to environment differences
between Travis and development machines. See this npm issue for some
background for why this change is needed:

https://github.com/npm/npm/issues/17722