fix test broken on recent versions of enzyme
When running against versions of enzyme > 3.3.0, one of the frontend
tests fails. The test actually had an error: it was erroneously trying
to select a <button class=".o-expandable_cue"> using a
".o-expandable_cue button" selector. This worked in enzyme <= 3.3.0 but
was fixed in 3.4.0, in https://github.com/airbnb/enzyme/pull/1680.

See this comment for someone else encountering the same thing:

https://github.com/airbnb/enzyme/pull/1680#issuecomment-414838829

To test, disable the use of package-lock with "npm config set
package-lock false", then run "npm install", then "npm test". This will
fail without the test to the change and pass with this change.

Thanks to @wpears for assistance with this.