remove version from package/package-lock.json
This package isn't going to be published to npm, so the version field is
optional. From https://docs.npmjs.com/files/package.json:

> If you plan to publish your package, the most important things in your
package.json are the name and version fields as they will be required.
The name and version together form an identifier that is assumed to be
completely unique. Changes to the package should come along with changes
to the version. If you don't plan to publish your package, the name and
version fields are optional.