properly fix frontend build and wheel names
The frontend build script run as part of pip install should do both npm
install and run the build, so that someone running it locally from
scratch will get a working build.

Additionally, on Travis, now that the build is run as part of tox (which
runs pip install), the package-lock.json file could be modified later in
the process. So the reset of that file needs to be moved down to just
before the wheel is built.