Generate complaint path URLs with clean IDs
We have a function to strip HTML tags like <em> out of complaint IDs
before rendering them. It was erroneously being applied only to the rendered
ID numbers and not also the URLs that get generated from the ID numbers.

To see the bug in action, try clicking on the complaint that appears in
these search results:

https://www.consumerfinance.gov/data-research/consumer-complaints/search/?date_received_max=2023-10-12&date_received_min=2020-10-12&page=1&searchField=all&searchText=4340364&size=25&sort=created_date_desc&tab=List

See DATAP-1210